Lab5/pzpi-22-6-skriahin-oleh-lab5/Back/SmartShield/incidents/serializers.py
from rest_framework import serializers
from .models import IncidentReport

class IncidentReportSerializer(serializers.ModelSerializer):
    class Meta:
        model = IncidentReport
        fields = '__all__'
        read_only_fields = ['created_at']