Lab5/pzpi-22-6-skriahin-oleh-lab5/Back/SmartShield/requirements.txt
asgiref==3.8.1
attrs==24.2.0
autobahn==24.4.2
Automat==25.4.16
CacheControl==0.14.3
cachetools==5.5.2
certifi==2024.8.30
cffi==1.17.1
channels==4.2.2
channels_redis==4.2.1
charset-normalizer==3.4.0
constantly==23.10.4
cryptography==44.0.0
daphne==4.2.0
defusedxml==0.8.0rc2
dj-database-url==2.3.0
Django==5.1.3
django-cors-headers==4.7.0
django-extensions==4.1
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djoser==2.3.1
drf-spectacular==0.28.0
drf-yasg==1.21.8
firebase-admin==6.8.0
google-api-core==2.24.2
google-api-python-client==2.170.0
google-auth==2.40.2
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-firestore==2.20.2
google-cloud-storage==3.1.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpcio==1.71.0
grpcio-status==1.71.0
gunicorn==23.0.0
httplib2==0.22.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
msgpack==1.1.0
oauthlib==3.2.2
packaging==24.2
proto-plus==1.26.1
protobuf==5.29.4
psycopg2-binary
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyJWT==2.10.1
pyOpenSSL==25.1.0
pyparsing==3.2.3
python3-openid==3.2.0
pytz==2024.2
PyYAML==6.0.2
redis==6.1.0
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.22.3
rsa==4.9.1
service-identity==24.2.0
setuptools==80.8.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.2
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
whitenoise==6.8.2
zope.interface==7.2
